% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loads data}
\usage{
load_data(rel_path, remote = FALSE)
}
\arguments{
\item{rel_path}{The experiment and data you want to load (e.g., snitz_2013/behavior.csv)}

\item{remote}{Load data locally FALSE or remomtely from pyrfume GitHub (TRUE)}
}
\value{
A data frame
}
\description{
Loads specified experimental data
}
\examples{
load_data("snitz_2013/behavior.csv")
}
